// src/types/generated/Animal.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AnimalSize } from "./AnimalSize";
import type { Motion } from "./Motion";
import type { Species } from "./Species";

export type Animal = { id: string, label: string, species: Species, size: AnimalSize, regionId: string | null, motion: Motion, };