// src/types/generated/SpeciesProfile.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Species } from "./Species";

export type SpeciesProfile = { species: Species, baseSizePx: number, movementFootprintPx: number, walkSpeedPxPerSecond: number, friction: number, minimumSpeedFraction: number, maximumTossSpeedMultiplier: number, turnMinMs: number, turnMaxMs: number, frameIntervalMs: number, };